+++ util/ibc/ibc_test.go
@@ -1,0 +1,49 @@
+package ibc
+
+import (
+	"strconv"
+	"strings"
+	"testing"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	ibctransfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
+	"github.com/tendermint/tendermint/crypto"
+	"gotest.tools/v3/assert"
+)
+
+func TestGetFundsFromPacket(t *testing.T) {
+	denom := strings.Join([]string{
+		"transfer",
+		"dest_chain",
+		"quark",
+	}, "/")
+
+	amount := strconv.Itoa(1)
+	data := ibctransfertypes.NewFungibleTokenPacketData(
+		denom,
+		amount,
+		AddressFromString("a3"),
+		AddressFromString("a4"),
+		"memo",
+	)
+
+	famount, fdenom, err := GetFundsFromPacket(data.GetBytes())
+	assert.NilError(t, err)
+	assert.Equal(t, denom, fdenom)
+	assert.Equal(t, famount.String(), amount)
+}
+
+func TestGetLocalDenom(t *testing.T) {
+	denom := strings.Join([]string{
+		"transfer",
+		"dest_chain",
+		"quark",
+	}, "/")
+
+	rdenom := GetLocalDenom(denom)
+	assert.Equal(t, rdenom, denom)
+}
+
+func AddressFromString(address string) string {
+	return sdk.AccAddress(crypto.AddressHash([]byte(address))).String()
+}